Create  linked list to implement an image viewer with previous and next buttons:
linked list to store the images, with the tail of the list representing the current image. The addImage function adds a new image to the end of the list, the getCurrentImage function returns the current image, and the prevImage and nextImage functions move the current image to the previous or next image in the list, respectively. The program reads in a list of images from the user, and then allows the user to navigate through the images using the previous and next buttons.