Create implementation of  Stack  to Scratch card’s earned after Google pay transaction
This program uses a stack data structure to store the scratch card rewards. The rewards stack is initialized empty, and then rewards are read in from the user and added to the top of the stack using the emplace function. The addReward function simply adds the reward to the stack and prints a message indicating that the reward has been added. The redeemRewards function takes the stack of rewards and a map of reward values as input, and then pops the rewards off the top of the stack one by one, calculates the total reward amount using the reward values from the map, and prints the total reward amount. If a reward type is not found in the map, it is considered invalid and is not included in the total reward amount. This program allows the user to add multiple rewards and then redeem them all at once, with the latest rewards appearing first when they are redeemed.