Create complete code for the circular linked list implementation in C++ that you can use to switch between applications and programs in an operating system:
This code will insert four applications/programs into the circular linked list and then display the current application/program and the list of applications/programs. It will then switch to the next application/program and display the current application/program and the list of applications/programs again. Finally, it will switch to the previous application/program and display the current application/program and the list of applications/programs once more.