Create C++ code that you can use to create a media playlist and play the previous and next song:
This code will insert four songs into the circular linked list and then display the current song and the playlist. It will then play the next song and display the current song and the playlist again. Finally, it will play the previous song and display the current song and the playlist once more.