Create divide and conquer algorithm Merge sort
mergeSort function which takes a vector and the range of the array that needs to be sorted. It then divides the array in half and recursively sorts each half. The merge function is then used to merge the two sorted halves back together in sorted order. The main function provides an example of how to use the mergeSort function, by creating a vector of integers and then calling the mergeSort function on it.